pyproject.toml
[project]
name="t-route"
authors = [
    {name = "DongHa Kim", email = "dongha.kim@noaa.gov"},
    {name = "Sean Horvath", email = "sean.horvath@noaa.gov"},
    {name = "Amin Torabi", email = "amin.torabi@noaa.gov"},
    {name = "Zach Jurgen", email = "jurgen.zach@noaa.gov"},
]
dynamic = ["version"]
dependencies = [
    "pip>=22.1.0",
    "setuptools>=64.0,<69.0",
    "wheel",
    "numpy~=1.0",
    "Cython>3,!=3.0.4",
    "pandas >=1.1.5, <=2.2.0",
    "xarray",
    "netcdf4",
    "pyyaml",
    "toolz",
    "joblib",
    "deprecated",
    "pyarrow<12.0",
    "geopandas",
    "pydantic<2",
]


[project.optional-dependencies]
doc = [
    "pydata-sphinx-theme[doc]==0.15.4",
    "sphinx-mdinclude==0.6.2",
]
test = [
    "pytest==8.3.2",
]
